atualizado

+++ Views1/NovoProduto.xaml.cs
@@ -1,3 +1,4 @@
+
 using MauiAppMinhasCompras.Models;
 using System.Threading.Tasks;
 
@@ -5,29 +6,30 @@ namespace MauiAppMinhasCompras.Views1;
 
 public partial class NovoProduto : ContentPage
 {
-	public NovoProduto()
-	{
-		InitializeComponent();
-	}
+    public NovoProduto()
+    {
+        InitializeComponent();
+    }
 
     private async void ToolbarItem_Clicked(object sender, EventArgs e)
     {
-		try
-		{
-			Produto p = new Produto
-			{
-				Descricao = txt_descricao.Text,
-				Quantidade = (int)Convert.ToDouble(txt_quantidade.Text),
-				Preco = (decimal)Convert.ToDouble(txt_preco.Text)
+        try
+        {
+            Produto p = new Produto
+            {
+                Descricao = txt_descricao.Text,
+                Quantidade = Convert.ToDouble(txt_quantidade.Text),
+                Preco = Convert.ToDouble(txt_preco.Text)
 
             };
 
-			await App.Db.Insert(p);
-			await DisplayAlert("Sucesso", "Registro Inserido", "OK");
+            await App.Db.Insert(p);
+            await DisplayAlert("Sucesso", "Registro Inserido", "OK");
 
-		} catch(Exception ex)
-		{
-			await DisplayAlert("Ops", ex.Message, "OK");
-		}
+        }
+        catch (Exception ex)
+        {
+            await DisplayAlert("Ops", ex.Message, "OK");
+        }
     }
 }
